% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCmat_for_BigShareDim.R
\name{GetCmat_for_BigShareDim}
\alias{GetCmat_for_BigShareDim}
\title{Compute RV from a list with big matrices}
\usage{
GetCmat_for_BigShareDim(data.list)
}
\arguments{
\item{data.list}{a list of matrices}
}
\value{
an RV matrix that stores the pair-wise RV coefficients of tables in the data.list
}
\description{
Compute RV from a list with big matrices
}
