% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFAnormMat.R
\name{MFAnormMat}
\alias{MFAnormMat}
\title{MFAnormMat}
\usage{
MFAnormMat(Y, BigShareDim = FALSE)
}
\arguments{
\item{Y}{A matrix to be MFA-normalized}

\item{BigShareDim}{FALSE (default). whether its a high dimensional matrix that requires regularized svd}
}
\value{
An MFA-normalized Y
}
\description{
MFAnormMat
}
\examples{
## a random 5 x 6 matrix
X <- matrix(rnorm(30), nrow = 5)
MFAnormMat(X)
}
