% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmds.new.R
\name{mmds.new}
\alias{mmds.new}
\title{A new version metric mulidimensional scaling}
\usage{
mmds.new(
  DistanceMatrix,
  masses = NULL,
  is.distance = TRUE,
  double.center = TRUE
)
}
\arguments{
\item{DistanceMatrix}{a distance matrix}

\item{masses}{NULL (default). Weights for each observations of the distance matrix.}

\item{is.distance}{If TRUE (default), the matrix will be processed as a distance matrix; if not, it will be processed as a covariance matrix}

\item{double.center}{If is.distance = FALSE, this is an option of deciding whether to double-center the matrix.}
}
\description{
A new version metric mulidimensional scaling
}
