% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RV4BigMatrix.R
\name{RV4BigMatrix}
\alias{RV4BigMatrix}
\title{Compute RV for two big matrices (with many rows)}
\usage{
RV4BigMatrix(Amat, Bmat)
}
\arguments{
\item{Amat}{a matrix with matching number of columns with Bmat}

\item{Bmat}{a matrix with matching number of columns with Amat}
}
\value{
an RV coefficient between Amat and Bmat with matching columns
}
\description{
Compute RV for two big matrices (with many rows)
}
