% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightCbind.R
\name{WeightCbind}
\alias{WeightCbind}
\title{Weighting the tables in a list and concatenating them by columns}
\usage{
WeightCbind(List, weight)
}
\arguments{
\item{List}{A list of K data matrices with matching rows}

\item{weight}{K weights, one for each table}
}
\value{
A big matrix with all weighted tables concatenated by columns
}
\description{
Weighting the tables in a list and concatenating them by columns
}
