% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wines2012List.R
\docType{data}
\name{wines2012List}
\alias{wines2012List}
\title{Wine sensory dataset (2012)}
\format{
A list of matrices representing wine sensory tables.
\describe{
\item{wines2012List}{A list containing sensory evaluation matrices for wines.}
}
}
\source{
Example dataset used in STATIS demonstrations
}
\usage{
wines2012List
}
\description{
Example dataset used to demonstrate STATIS analysis.
}
\keyword{datasets}
